Renamed a bunch of mixin file names for pluralization consistency.
