+++ src/hi/apps/console/console_mixins.py
