+++ src/hi/apps/entity/view_mixins.py
