+++ src/hi/apps/sense/view_mixins.py
