+++ src/hi/apps/sense/views.py
@@ -7,7 +7,7 @@ from hi.apps.common.pagination import compute_pagination_from_queryset
 from hi.hi_async_view import HiModalView
 
 from .models import SensorHistory
-from .view_mixin import SenseViewMixin
+from .view_mixins import SenseViewMixin
 
 logger = logging.getLogger(__name__)
